api/pyproject.toml
[tool.poetry]
name = "treediculous"
version = "0.4.0"
description = ""
authors = ["patacoing <romain.lefebvre1104@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.26.2"
pillow = "^11.0.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
opencv-python-headless = "^4.10.0.84"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
